\name{tipsntRicks-package}
\alias{tipsntRicks-package}
\alias{tipsntRicks}
\docType{package}

\title{
  A collection of R tips and tricks
}


\description{
Code that comes with the 'tipsntRicks - a collection of R tips and tricks' website.
}


\author{
Authors of the respective functions are cited in the respective man pages.

Maintainer: Laurent Gatto <lg390@cam.ac.uk>
}

\keyword{ package }
